requirements.txt
pytorch-lightning==2.4.0
einops
plotly
omegaconf
trimesh
rembg
gradio==5.5.0
huggingface_hub[cli]==0.26.2
transformers==4.46.1
tokenizers==0.20.1
diffusers==0.31.0
accelerate==1.0.1
xformers==0.0.22.post7
grpcio-status==1.33.2 
protobuf==3.20.2
safetensors
imageio[ffmpeg]
kornia 
timm
plyfile
open3d==0.18.0
xatlas
fpsample
fast-simplification
numpy==1.26.1
git+https://github.com/ashawkey/diff-gaussian-rasterization.git
git+https://github.com/hbb1/diff-surfel-rasterization.git
git+https://github.com/NVlabs/nvdiffrast.git
git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8